default to first character as start

+++ xc/programs/xfd/gridP.h
@@ -20,7 +20,7 @@ typedef struct _FontGridPart {
     Pixel		box_pixel;	/* for box_chars */
     XtCallbackList	callbacks;	/* for notifying caller */
     int			internal_pad;	/* extra padding inside grid */
-    unsigned int	start_char;	/* first character of grid */
+    long		start_char;	/* first character of grid */
     int			grid_width;	/* width of grid lines */
     /* private data */
     GC			text_gc;	/* printing text */
